feat: Added Service/Router separation for [POST] /api/v1/git_tokens/
Summary:

- Split the router into a router layer and a service layer.
- abstracted the database communication between the service layer and the Database Model.
