+++ devdox/app/exceptions/exception_constants.py
@@ -2,6 +2,7 @@ AUTH_FAILED = "Access denied. Please log in again."
 INVALID_BEARER_TOKEN_SCHEMA = "Missing or malformed Authorization header"
 SERVICE_UNAVAILABLE = "Service temporarily unavailable. Please try again later."
 GENERIC_BAD_REQUEST = "Bad Request"
+TOKEN_MISSING = "Token is missing"
 GENERIC_RESOURCE_NOT_FOUND = "Resource Not Found"
 USER_RESOURCE_NOT_FOUND = "User Not Found"
 GIT_LABEL_TOKEN_RESOURCE_NOT_FOUND = "Git Label Token Not Found"
